+++ tests/TestValkeyGlide.php
@@ -98,6 +98,8 @@ require_once __DIR__ . "/ValkeyGlideClusterJsonTest.php";
 require_once __DIR__ . "/ValkeyGlideSearchTest.php";
 require_once __DIR__ . "/ValkeyGlideClusterSearchTest.php";
 require_once __DIR__ . "/ClientSideCacheTest.php";
+require_once __DIR__ . "/AddressResolverTest.php";
+require_once __DIR__ . "/AddressResolverClusterTest.php";
 
 function getClassArray($classes)
 {
@@ -134,7 +136,9 @@ function getTestClass($class)
         'valkeyglideclusterjson' => 'ValkeyGlideClusterJsonTest',
         'valkeyglidesearch' => 'ValkeyGlideSearchTest',
         'valkeyglideclustersearch' => 'ValkeyGlideClusterSearchTest',
-        'clientsidecache' => 'ClientSideCacheTest'
+        'clientsidecache' => 'ClientSideCacheTest',
+        'addressresolver' => 'AddressResolverTest',
+        'addressresolvercluster' => 'AddressResolverClusterTest',
     ];
 
     /* Return early if the class is one of our built-in ones */
@@ -181,6 +185,8 @@ $default_classes = implode(',', [
     'valkeyglidedns',
     'valkeyglideclusterdns',
     'iamauth',
+    'addressresolver',
+    'addressresolvercluster',
 ]);
 
 $classes = getClassArray($opt['class'] ?? $default_classes);